IQ save option, footer interpretation and file saving adjustments
Added self.miq_flag to allow user to choose to perform digital down
conversion and save a .mat file or not. Speeds up processing
significantly if you only want header and/or footer information.
Removed the frame description field from parse_footer since it contains
proprietary and undocumented information that is not useful to the user.
changed formatting of the frame_status field so that it prints in
binary, which is more useful. Added formatting to save_footer_file,
including tabs between columns and labels for each column.